Updates
Added a visual canvas on the player to show which lens you are on.
Added the Raycast script to the player prefab.
Moved the raycast input map to the main input map.
Set up the reverse lens, with a temporary material. Reverse lens will activate/deactivate all tagged items on entering/exiting reverse and tagging/untagging in reverse lens.
Added Tags/scripts to items in the tutorial scene